Convert all route partials into route extensions, then realise I didnt need to. Ah well.
