+++ plugins/ShinyAccess/config/routes.rb
@@ -6,16 +6,11 @@
 #
 # ShinyCMS is free software; you can redistribute it and/or modify it under the terms of the GPL (version 2 or later)
 
-# Routes for ShinyAccess plugin
-
-require_relative '../../../plugins/ShinyCMS/lib/import_routes'
-
 ShinyAccess::Engine.routes.draw do
   scope format: false do
     # Admin area
     scope path: :admin, module: :admin do
-      # with_paging and with_search
-      import_routes partial: :admin_route_concerns
+      extend ShinyCMS::Routes::AdminConcerns  # with_paging and with_search
 
       scope path: :access do
         resources :groups, except: %i[ index show ], concerns: %i[ with_paging with_search ] do
