+++ plugins/ShinyBlog/config/routes.rb
@@ -6,8 +6,6 @@
 #
 # ShinyCMS is free software; you can redistribute it and/or modify it under the terms of the GPL (version 2 or later)
 
-require_relative '../../../plugins/ShinyCMS/lib/import_routes'
-
 ShinyBlog::Engine.routes.draw do
   scope format: false do
     yyyy_mm = { year: %r{\d\d\d\d}, month: %r{\d\d} }
@@ -22,8 +20,7 @@ ShinyBlog::Engine.routes.draw do
 
     # Admin area
     scope path: 'admin', module: 'admin' do
-      # with_paging and with_search
-      import_routes partial: :admin_route_concerns
+      extend ShinyCMS::Routes::AdminConcerns  # with_paging and with_search
 
       resources :blog_posts, path: 'blog', except: %i[ index show ], concerns: %i[ with_paging with_search ]
     end
