+++ plugins/ShinyCMS/config/inflections.yml
@@ -14,6 +14,7 @@ acronym:
   - IP
   - HTML
   - MJML
+  - PHP
   - SEO
   - ShinyCMS
   - URL
