+++ plugins/ShinyPages/config/routes.rb
@@ -6,8 +6,6 @@
 #
 # ShinyCMS is free software; you can redistribute it and/or modify it under the terms of the GPL (version 2 or later)
 
-require_relative '../../../plugins/ShinyCMS/lib/import_routes'
-
 ShinyPages::Engine.routes.draw do
   scope format: false do
     # Main site
@@ -15,8 +13,7 @@ ShinyPages::Engine.routes.draw do
 
     # Admin area
     scope path: :admin, module: :admin do
-      # with_paging and with_search
-      import_routes partial: :admin_route_concerns
+      extend ShinyCMS::Routes::AdminConcerns  # with_paging and with_search
 
       concern :sortable do
         put :sort, on: :collection
