+++ plugins/ShinySEO/config/routes.rb
@@ -6,8 +6,6 @@
 #
 # ShinyCMS is free software; you can redistribute it and/or modify it under the terms of the GPL (version 2 or later)
 
-# Routes for ShinySEO plugin
-
 ShinySEO::Engine.routes.draw do
   scope format: false do
     # ...
